requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
biopython==1.85
bleach==6.3.0
CAI-PyPI==2.0.1
certifi==2025.11.12
charset-normalizer==3.4.3
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
dtw-python==1.7.2
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fastdtw==0.3.4
fastjsonschema==2.21.2
filelock==3.13.1
flatbuffers==25.9.23
fonttools==4.60.0
frozenlist==1.8.0
fsspec==2024.6.1
GridDataFormats==1.0.2
hf-xet==1.1.10
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.12.3
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lightning-utilities==0.15.2
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mda-xdrlib==0.2.0
MDAnalysis==2.9.0
mistune==3.2.0
mmtf-python==1.1.3
mpmath==1.3.0
mrcfile==1.5.4
msgpack==1.1.1
multidict==6.7.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.2
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.4
python_codon_tables==0.1.18
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rotary-embedding-torch==0.8.9
rpds-py==0.30.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
setuptools==78.1.1
six==1.17.0
soupsieve==2.8
stack_data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
torch_geometric==2.2.0
torchmetrics==1.8.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.45.1
widgetsnbextension==4.0.15
yarg==0.1.9
yarl==1.22.0
zipp==3.23.0
